minor code cleanup.
